+++ undergroundbot-v0/commands/admin/kick.js
@@ -23,23 +23,43 @@ module.exports = class SayCommand extends Command {
       return msg.reply(
         "you must mention someone or not add any extra arguments!"
       );
-    adminCheck(msg.author.id, function(result) {
-      if (result == true) {
-        adminCheck(mentions.id, function(result) {
-          if (result == true || mentions.kickable == false) {
-            if (mentions.id == botsudoid) {
-              db.exec(`delete from admins where id=${msg.author.id}`);
+    cooldownCheck(msg.author.id, "kick", function(result) {
+      // check if our dear author is on a cooldown
+      let timeRemaining = result;
+      let timeRemainingFormatted = format(timeRemaining);
+      if (timeRemaining != 0)
+        return msg.reply(
+          `Slow down pal! Remaining cooldown: \`${timeRemainingFormatted}\``
+        );
+      adminCheck(msg.author.id, function(result) {
+        if (result == true) {
+          adminCheck(mentions.id, function(result) {
+            if (result == true || mentions.kickable == false) {
+              if (mentions.id == botsudoid) {
+                db.exec(`delete from admins where id=${msg.author.id}`);
+                return msg.reply(
+                  "You can't do that. You have been removed from the admin list because you are mean."
+                );
+              }
+              return msg.reply("thats an admin doofus, i cant do that");
+            } else {
+              // oh good! author wasnt on cooldown, now he is
+              let cooldownTime = Date.now() + 1800000;
+
+              // insert into cooldowns table
+              db.exec(
+                `insert into cooldowns values (${
+                  msg.author.id
+                }, 'kick', ${cooldownTime})`
+              );
+              mentions.kick();
               return msg.reply(
-                "You can't do that. You have been removed from the admin list because you are mean."
+                `${mentions.username} was kicked by ${msg.author.username}`
               );
             }
-            return msg.reply("thats an admin doofus, i cant do that");
-          } else {
-            mentions.kick()
-            return msg.reply(`${mentions.username} was kicked by ${msg.author.username}`);
-          }
-        });
-      } else return msg.reply("You are not a bot admin.");
+          });
+        } else return msg.reply("You are not a bot admin.");
+      });
     });
   }
 };
